+++ CHANGELOG.md
@@ -6,6 +6,7 @@ omnibus-gitlab repository.
 11.6.0
 
 - Upgrade Ruby to 2.5.3 !2806
+- Move env directory of all services to `/opt/gitlab/etc/<service>/env` !2825
 
 11.5.0